First commit: INO executable file
The code enables to use the lepton 2.5 module to take images of frames of 60 by 80 pixels. Use Arduino IDE serial plotter to get these values.

Each image is taken at an interval of 2 seconds. Each image is segregated with string "Start" and "End".

The post processing has to be done to convert the data into meaningful thermal image.